test: write 2 sector patterns, verify exact disk contents
In each of our test programs, the first write uses the first
pattern and the second write uses the second pattern. This
allows us to check that the second write is capable of
overwriting the first.

Extend the verification-by-streaming step to check the
pattern referenced by a 'Copy' request.

This test currently passes.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>