Remove the 'handle' wrapper over file_descr-like things
The only function of this 'handle' was to prevent the accidental
double-closing of Unix file descriptors. Since this is really
a Unix-ism, remove from the generic code. A future version of the
Vhd_lwt module can do this instead.

Note this exposed a file descriptor handling buf: if one opens
a vhd and then creates a snapshot, one ends up with two references
to the original vhd and it's not clear when this should be closed.
The current code doesn't call close() recursively so it never
double-closes but will occasionally leak.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>