CP-13838: only include Vhd blocks if a predicate `include_block` is true
Previously we assumed that we would not be able to figure out if
raw-formatted data is zero or not. However it is possible on some systems
to use `SEEK_DATA` and `SEEK_HOLE`, so this patch refactors the code
to only include a block if the predicate `include_block` is true.

Signed-off-by: David Scott <dave.scott@citrix.com>