Limit number of sectors to coalesce
We limit the number of sectors to coalesce when generating a list of
elements to be copied. The limit is currently set to 4 MiB. This allows
the progress bar to be displayed at a finer granularity rather than
going from 0 to 100 in one go due to one big sector being generated.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>